1.0 Malo Mart Music Replacement
This is the first upload of the files. 1.0.0 replaces the Shop Theme Music in the game REPO with Malo Mart's theme music from the game Legend of Zelda: Twilight Princess.
I'm new to git hub so I am unsure if I'm doing things right. I wanted to have Malo Mart Music from Legend of Zelda: Twilight Princess in the game Repo. But the two mods that I found that add that weren't working once loaded into the game. I looked at other files from shop music replacement mods that worked for me and attempted to recreate that with the Malo Mart music instead. I just wanted to upload it to Thunderstore so I could throw the mod into a multiplayer file for friends, which has lead me to GitHub. Hopefully it works and I did things correctly.